🧹 Rename temp screen folder back to "scr/screens"
