+++ src/screens/StopScreen.js
